Add build and test instructions
* Don't build on openbsd
* Don't test fuse on travis
