+++ cmd/restic/cmd_mount.go
@@ -1,3 +1,5 @@
+// +build !openbsd
+
 package main
 
 import (
