+++ test/backend.go
@@ -18,6 +18,7 @@ var (
 	TestCleanup           = getBoolVar("RESTIC_TEST_CLEANUP", true)
 	TestTempDir           = getStringVar("RESTIC_TEST_TMPDIR", "")
 	RunIntegrationTest    = getBoolVar("RESTIC_TEST_INTEGRATION", true)
+	RunFuseTest           = getBoolVar("RESTIC_TEST_FUSE", true)
 	TestSFTPPath          = getStringVar("RESTIC_TEST_SFTPPATH", "/usr/lib/ssh:/usr/lib/openssh")
 	TestWalkerPath        = getStringVar("RESTIC_TEST_PATH", ".")
 	BenchArchiveDirectory = getStringVar("RESTIC_BENCH_DIR", ".")
